+++ conn.go
@@ -1,0 +1,41 @@
+package sm_yamux
+
+import (
+	"github.com/libp2p/go-libp2p-core/mux"
+	"github.com/libp2p/go-yamux"
+)
+
+// conn implements mux.MuxedConn over yamux.Session.
+type conn yamux.Session
+
+// Close closes underlying yamux
+func (c *conn) Close() error {
+	return c.yamux().Close()
+}
+
+// IsClosed checks if yamux.Session is in closed state.
+func (c *conn) IsClosed() bool {
+	return c.yamux().IsClosed()
+}
+
+// OpenStream creates a new stream.
+func (c *conn) OpenStream() (mux.MuxedStream, error) {
+	s, err := c.yamux().OpenStream()
+	if err != nil {
+		return nil, err
+	}
+
+	return (*stream)(s), nil
+}
+
+// AcceptStream accepts a stream opened by the other side.
+func (c *conn) AcceptStream() (mux.MuxedStream, error) {
+	s, err := c.yamux().AcceptStream()
+	return (*stream)(s), err
+}
+
+func (c *conn) yamux() *yamux.Session {
+	return (*yamux.Session)(c)
+}
+
+var _ mux.MuxedConn = &conn{}
